code-challenges/1nfluencers/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "allainclair", email = "allainclair@gmail.com"},
]
dependencies = [
    "python-dotenv>=0.20.0",
    "requests>=2.28.1",
    "selenium>=4.4.0",
    "webdriver-manager>=3.8.3",
    "scrapy>=2.6.2",
]
requires-python = ">=3.10"
license = {text = "MIT"}
[project.optional-dependencies]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.1.2",
    "icecream>=2.1.3",
]
